Refactor: move wikidta models to data.models
