+++ app/src/main/java/fr/free/nrw/commons/mwapi/OkHttpJsonApiClient.java
@@ -7,7 +7,6 @@ import android.text.TextUtils;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import com.google.gson.Gson;
-import fr.free.nrw.commons.auth.SessionManager;
 import fr.free.nrw.commons.campaigns.CampaignResponseDTO;
 import fr.free.nrw.commons.explore.depictions.DepictsClient;
 import fr.free.nrw.commons.location.LatLng;
@@ -21,7 +20,7 @@ import fr.free.nrw.commons.profile.leaderboard.UpdateAvatarResponse;
 import fr.free.nrw.commons.upload.FileUtils;
 import fr.free.nrw.commons.upload.structure.depictions.DepictedItem;
 import fr.free.nrw.commons.utils.ConfigUtils;
-import fr.free.nrw.commons.wikidata.model.GetWikidataEditCountResponse;
+import fr.free.nrw.commons.data.models.wikidata.GetWikidataEditCountResponse;
 import io.reactivex.Observable;
 import io.reactivex.Single;
 import java.io.IOException;
