+++ app/src/main/java/fr/free/nrw/commons/wikidata/WikidataInterface.java
@@ -2,20 +2,15 @@ package fr.free.nrw.commons.wikidata;
 
 import androidx.annotation.NonNull;
 
-import com.google.gson.JsonObject;
 import org.wikipedia.dataclient.mwapi.MwQueryResponse;
 
-import fr.free.nrw.commons.wikidata.model.AddEditTagResponse;
-import fr.free.nrw.commons.wikidata.model.WbCreateClaimResponse;
+import fr.free.nrw.commons.data.models.wikidata.WbCreateClaimResponse;
 import io.reactivex.Observable;
-import okhttp3.RequestBody;
 import retrofit2.http.Field;
 import retrofit2.http.FormUrlEncoded;
 import retrofit2.http.GET;
 import retrofit2.http.Headers;
-import retrofit2.http.Multipart;
 import retrofit2.http.POST;
-import retrofit2.http.Part;
 
 import static org.wikipedia.dataclient.Service.MW_API_PREFIX;
