+++ app/src/test/kotlin/fr/free/nrw/commons/ModelFunctions.kt
@@ -7,7 +7,7 @@ import fr.free.nrw.commons.nearby.Label
 import fr.free.nrw.commons.nearby.Place
 import fr.free.nrw.commons.nearby.Sitelinks
 import fr.free.nrw.commons.upload.structure.depictions.DepictedItem
-import fr.free.nrw.commons.wikidata.model.DepictSearchItem
+import fr.free.nrw.commons.data.models.wikidata.DepictSearchItem
 import org.wikipedia.wikidata.*
 import java.util.*
